don't clobber a previously set level filter
There's a bit of a race with the `log` API for initializing because
it covers two atomic values. It's possible to observe a logger and
attempt to use it before the max level is set. This change doesn't
really make that race any racier than it was already, it just prevents
the surprising behaviour of clobbering a previously set max level
filter even if the logger itself isn't set.