[02 of 10]: filter::parse_spec(): add more parse_default* tests
Add tests that elaborate on the existing parse_default() unit
test. These new tests exercise all of log::Level variants (plus the
"OFF" pseudo log level), and also serve as minimal examples of how
to set the default log level. For example, they exercise both
lowercase and uppercase variants. There is one mixed case variant
test thrown in, too, since that is specifically allowed for by the
external log::Level::from_str() implementation on which we depend.

There is also a test to ensure all log::Level variants are accounted
for in the above tests. That one is intended to avoid the false
sense of comfort from passing tests in the unlikely event that a new
variant is added to the externally defined log::Level enum. (Note
that this compile-time check does not include the "OFF" pseudo log
level, as it exists outside of the log::Level enum.)