[06 of 10]: docs: clarify default enabled/disabled log level behavior
The "Enabling logging" section of the crate-level docs are here
updated to clarify the default behavior as it pertains to the
logging level.

Previously, the first paragraph stated that all logging is disabled
except for the default level, but the third paragraph stated that if
no log level was specified then /all/ logging was enabled.

It turns out that the third paragraph was really intended as a
continuation of the second paragraph, the applicable context of
which is limited to those scenarios in which a logging directive has
been supplied.

This changeset reworks the paragraph structure slightly to make the
intent more clear, and also to make it easy to quickly identify the
most important aspects:

    A. Old para 1 split into two separate paras:
        - first states clearly the default behavior (now in bold)
        - second introduces 'RUST_LOG' and logging directives

    B. Old para 3 split into two separate paras:
        - first now qualifies the statement with "When specifying
          the crate name or a module path..."
        - second introduces log level names

    C. The existing statement that describes the behavior when only
       a log level is provided (it sets the global log level) is
       now in bold markdown.

It might make sense to refine this further at some point so that the
flow of the text answers in order the user's questions:

    1. "What is the default behavior?"  (Already first; good.)

    2. "How do I set the log level for the entire app?"

    3. "How do I set the log level more surgically?"

Our current flow is: (1), (3), (2)